account for GPIO changes in bullseye
the sysfs interface for controlling GPIO is deprecated.
This change updates the amp power script to use the command line tools
from the gpiod package. One bit of lost functionality is the ability
to read the GPIO value without resetting it back to `input`. This means
if we tried to read the status, it would turn the amp off.